+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.5] --2025-09-21
+
+- Plugin 添加 ignoreEdited 参数，插件将会监听编辑信息事件
+- eatgif 利用 ffmpeg 将 gif 转成 webm 表情包
+- 添加 dotenv 依赖 以及 env 配置文件
+
 ## [0.2.4] --2025-09-13
 
 - 修复 Api.Message.deleteWithDelay 没有类型提醒